Add multiple variants of e2e templates
